feat(stdlib): enable moving average and cumulative sum optimizations by default
