+++ internal/feature/flags.go
@@ -173,30 +173,6 @@ func LabelPolymorphism() BoolFlag {
 	return labelPolymorphism
 }
 
-var optimizeMovingAverage = feature.MakeBoolFlag(
-	"Optimize Moving Average",
-	"optimizeMovingAverage",
-	"Jonathan Sternberg",
-	false,
-)
-
-// OptimizeMovingAverage - Enable an optimized version of moving average
-func OptimizeMovingAverage() BoolFlag {
-	return optimizeMovingAverage
-}
-
-var optimizeCumulativeSum = feature.MakeBoolFlag(
-	"Optimize Cumulative Sum",
-	"optimizeCumulativeSum",
-	"Jonathan Sternberg",
-	false,
-)
-
-// OptimizeCumulativeSum - Enable optimized version of cumulative sum
-func OptimizeCumulativeSum() BoolFlag {
-	return optimizeCumulativeSum
-}
-
 var optimizeSetTransformation = feature.MakeBoolFlag(
 	"Optimize Set Transformation",
 	"optimizeSetTransformation",
@@ -228,8 +204,6 @@ var all = []Flag{
 	vectorizeAddition,
 	vectorizeOperators,
 	labelPolymorphism,
-	optimizeMovingAverage,
-	optimizeCumulativeSum,
 	optimizeSetTransformation,
 }
 
@@ -247,8 +221,6 @@ var byKey = map[string]Flag{
 	"vectorizeAddition":                vectorizeAddition,
 	"vectorizeOperators":               vectorizeOperators,
 	"labelPolymorphism":                labelPolymorphism,
-	"optimizeMovingAverage":            optimizeMovingAverage,
-	"optimizeCumulativeSum":            optimizeCumulativeSum,
 	"optimizeSetTransformation":        optimizeSetTransformation,
 }
