+++ stdlib/universe/cumulative_sum_test.go
@@ -6,29 +6,9 @@ import (
 	"github.com/influxdata/flux"
 	"github.com/influxdata/flux/execute"
 	"github.com/influxdata/flux/execute/executetest"
-	"github.com/influxdata/flux/querytest"
+	"github.com/influxdata/flux/memory"
 	"github.com/influxdata/flux/stdlib/universe"
 )
-
-func TestCumulativeSumOperation_Marshaling(t *testing.T) {
-	data := []byte(`{"id":"cumulativeSum","kind":"cumulativeSum","spec":{}}`)
-	op := &flux.Operation{
-		ID:   "cumulativeSum",
-		Spec: &universe.CumulativeSumOpSpec{},
-	}
-	querytest.OperationMarshalingTestHelper(t, data, op)
-}
-
-func TestCumulativeSum_PassThrough(t *testing.T) {
-	executetest.TransformationPassThroughTestHelper(t, func(d execute.Dataset, c execute.TableBuilderCache) execute.Transformation {
-		s := universe.NewCumulativeSumTransformation(
-			d,
-			c,
-			&universe.CumulativeSumProcedureSpec{},
-		)
-		return s
-	})
-}
 
 func TestCumulativeSum_Process(t *testing.T) {
 	testCases := []struct {
@@ -261,13 +241,17 @@ func TestCumulativeSum_Process(t *testing.T) {
 	for _, tc := range testCases {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
-			executetest.ProcessTestHelper(
+			executetest.ProcessTestHelper2(
 				t,
 				tc.data,
 				tc.want,
 				nil,
-				func(d execute.Dataset, c execute.TableBuilderCache) execute.Transformation {
-					return universe.NewCumulativeSumTransformation(d, c, tc.spec)
+				func(id execute.DatasetID, alloc memory.Allocator) (execute.Transformation, execute.Dataset) {
+					tr, d, err := universe.NewCumulativeSumTransformation(id, tc.spec, alloc)
+					if err != nil {
+						t.Fatal(err)
+					}
+					return tr, d
 				},
 			)
 		})
